Initial version of TypeScript conversion finished
